+++ src/main/java/dao/UserDAOPostgreSQLImpl.java
@@ -40,7 +40,7 @@ public class UserDAOPostgreSQLImpl implements UserDAO{
     }
 
     @Override
-    public User selectUser(int id) throws SQLException {
+    public User selectUser(long id) throws SQLException {
         PreparedStatement statement = connection.prepareStatement("SELECT * FROM users WHERE id = ?");
         statement.setInt(1, id);
         ResultSet resultSet = statement.executeQuery();
@@ -72,9 +72,9 @@ public class UserDAOPostgreSQLImpl implements UserDAO{
 
 
     @Override
-    public boolean deleteUser(int id) throws SQLException {
+    public boolean deleteUser(long id) throws SQLException {
         PreparedStatement statement = connection.prepareStatement("DELETE FROM users WHERE id = ?");
-        statement.setInt(1, id);
+        statement.setLong(1, id);
         int affectedRows = statement.executeUpdate();
         return affectedRows > 0;
     }
